backend/requirements.txt
Django==4.2.*
djangorestframework==3.15.*
django-cors-headers==4.3.*
requests==2.31.*
apscheduler==3.10.*
